Fix loss weights initialization and add script to convert old embeddings to gensim keyed vectors

+++ deepbond/models/cnn.py
@@ -22,7 +22,7 @@ class CNN(Model):
 
     def build(self, options, loss_weights=None):
         if loss_weights is not None:
-            loss_weights = torch.tensor(loss_weights)
+            loss_weights = torch.tensor(loss_weights).float()
 
         word_embeddings = None
         if self.words_field.vocab.vectors is not None:
