+++ deepbond/models/rcnn.py
@@ -35,7 +35,7 @@ class RCNN(Model):
         # caps_embeddings_size = options.caps_embeddings_size
         hidden_size = options.hidden_size[0]
         if loss_weights is not None:
-            loss_weights = torch.tensor(loss_weights)
+            loss_weights = torch.tensor(loss_weights).float()
 
         word_embeddings = None
         if self.words_field.vocab.vectors is not None:
